Add workaround for kernel function arguments with side-effects.
hipLaunchKernel first locks the stream to provide thread-safe access,
and then invokes the host wrapper to launc the requested kernel.

If the arguments to the kernel are functions, and those other
functions call hip APIs which also equire access to the stream
(such as hipMemcpy), then deadlock ensues.  Workaround is to save
the arguments which call functions into temporary variables.
This allows them to call hip APIs before the stream is locked.

This workaround is needed in several other places as well.