Workaround to initialize a static class member variable within applications
This commit introduces an workaround to make hipcaffe be built with ToT HCC.
It changes include/caffe/data_reader.hpp so an static member variable:

caffe::DataReader::bodies_

is initialized in caffe applications, instead of libcaffe.so

The initialization logic is placed in include/caffe/data_reader_hack.inl .
Caffe applications are changed to include this file.

This should only be considered as an workaround. Proper fix in HCC which
handles initialization of static member variables in shared libraries shall
be devised.