[MLOPEN] First compilable commit
This is the first commit in hip-tot_hcc-mlopen branch, which aims to integrate
MLOpen kernels into hipcaffe, and have the system be built with ToT HCC.

To prevent too much code changes up front, the approach taken is to hid MLOpen
integration logic inside CUDNN helper classes in hipcaffe. So hipcaffe still
thinks it's running on top of CUDNN, just the implementation would be switched
to MLOpen.

- Introduce USE_ACCELERATED_NN switch, which turns on GPU support for hipcaffe
- Introduce USE_MLOPEN switch, which enables MLOpen integration to hipcaffe
- Changed semantic between USE_ACCELERATED_NN, USE_CUDNN, USE_MLOPEN so that:
  - USE_ACCELERATED_NN is the top-level switch to turn on GPU support
  - Use one of USE_MLOPEN or USE_CUDNN to pick the actual GPU path
- Changed headers / sources / kernels wrappers for CUDNN, so they can now be
  integrated with MLOpen. The integration logic are largely dummy at this
  moment.

In the future, when MLOpen integration is getting matured, it may be worthwhile
to refactor this architecture and:

- Separate logic between USE_CUDNN and USE_MLOPEN into different files.
- Change caffe.proto to add a new Engine "MLOPEN".
- Change layer_factory.cpp to properly set MLOpen as a new Engine.